test(#1): move tests outside pages directory
