+++ tests/__tests__/index.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 
-import Home from '../index';
+import Home from '../../pages';
 import '@testing-library/jest-dom';
 
 describe('Home', () => {
