+++ tests/__tests__/snapshot.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 
-import Home from '../index';
+import Home from '../../pages';
 
 it('renders homepage unchanged', () => {
   const { container } = render(<Home />);
